Move batch inverse to new module
