+++ math/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod batch_inverse;
 pub mod cyclic_group;
 pub mod elliptic_curve;
 pub mod errors;
